refactor: 타입 순환 린트 오류 해결
