+++ src/pages/HiresEdit/HiresEdit.tsx
@@ -3,25 +3,7 @@ import { useParams, useLocation } from 'react-router-dom';
 
 import { HiresCreate } from '..';
 import { previousHiresInfo } from '../HiresCreate/HiresCreate';
-
-export interface hireItemType {
-  postId?: number | undefined;
-  teamLogo: string | undefined;
-  date: string | undefined;
-  startTime: string | undefined;
-  endTime: string | undefined;
-  city: string | undefined;
-  region: string | undefined;
-  groundName: string | undefined;
-  position: string | undefined;
-  ageGroup: string | undefined;
-  teamMannerTemperature: number | undefined;
-  hiredPlayerNumber?: number | undefined;
-  hirePlayerNumber: number | undefined;
-  teamName: string | undefined;
-  teamManagerName: string | undefined;
-  detail: string | undefined;
-}
+import { hireItemType } from '@/types';
 
 interface hireItemWrapperType {
   hireItem: hireItemType;
