+++ src/types/Hires/Hires.ts
@@ -1,0 +1,18 @@
+export interface hireItemType {
+  postId?: number | undefined;
+  teamLogo: string | undefined;
+  date: string | undefined;
+  startTime: string | undefined;
+  endTime: string | undefined;
+  city: string | undefined;
+  region: string | undefined;
+  groundName: string | undefined;
+  position: string | undefined;
+  ageGroup: string | undefined;
+  teamMannerTemperature: number | undefined;
+  hiredPlayerNumber?: number | undefined;
+  hirePlayerNumber: number | undefined;
+  teamName: string | undefined;
+  teamManagerName: string | undefined;
+  detail: string | undefined;
+}
