+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './Team/Team';
 export * from './Match/Match';
+export * from './Hires/Hires';
